unit11/requirements.txt
# Core dependencies
llama-index>=0.12.0
llama-index-core>=0.12.0
llama-index-llms-openai>=0.3.0
llama-index-embeddings-openai>=0.3.0
llama-index-readers-web>=0.3.0
llama-index-agent-openai>=0.3.0

# Data processing
pandas>=2.0.0
pandas-ta>=0.3.0
numpy>=1.24.0

# Web framework
streamlit>=1.30.0

# API and data collection
requests>=2.31.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0 